pyproject.toml
[project]
name = "adit-radis-shared"
description = "Shared Django apps between ADIT and RADIS"
authors = [{ name = "Kai Schlamp" }]
license = "AGPL-3.0-or-later"
version = "0.0.0"
readme = "README.md"
requires-python = ">=3.12,<4.0"
dependencies = [
    "channels>=4.2.0",
    "crispy-bootstrap5>=2024.10",
    "cryptography>=44.0.1",
    "daphne>=4.1.2",
    "Django>=5.1.6",
    "django-block-fragments>=0.1.1",
    "django-cotton>=1.6.0",
    "django-crispy-forms>=2.3",
    "django-dbbackup>=4.2.1",
    "django-extensions>=3.2.3",
    "django-filter>=25.1",
    "django-htmx>=1.22.0",
    "django-loginas>=0.3.11",
    "django-registration-redux>=2.13",
    "django-revproxy>=0.13.0",
    "django-tables2>=2.7.5",
    "djangorestframework>=3.15.2",
    "dunamai>=1.23.0",
    "environs[django]>=14.1.1",
    "procrastinate[django]>=3.0.2",
    "psycopg[binary]>=3.2.5",
    "python-dotenv>=1.0.1",
    "requests>=2.32.3",
    "Twisted[tls,http2]>=24.11.0",
    "wait-for-it>=2.3.0",
    "watchfiles>=1.0.4",
    "whitenoise>=6.9.0",
]

[dependency-groups]
dev = [
    "argcomplete>=3.6.1",
    "debugpy>=1.8.12",
    "django-browser-reload>=1.18.0",
    "django-debug-permissions>=1.0.0",
    "django-debug-toolbar>=5.0.1",
    "django-stubs>=5.1.3",
    "django-test-migrations>=1.4.0",
    "djangorestframework-stubs>=3.15.3",
    "djlint>=1.36.4",
    "factory-boy>=3.3.2,<3.3.3",
    "Faker>=36.1.1",
    "ipykernel>=6.29.5",
    "ipython>=8.32.0",
    "pyright>=1.1.394",
    "pytest>=8.3.4",
    "pytest-asyncio>=0.25.3",
    "pytest-cov>=6.0.0",
    "pytest-django>=4.10.0",
    "pytest-mock>=3.14.0",
    "pytest-only>=2.1.2",
    "pytest-order>=1.3.0",
    "pytest-picked>=0.5.1",
    "pytest-playwright>=0.7.0",
    "pytest-timeout>=2.3.1",
    "pytest-watch>=4.2.0",
    "pywatchman>=2.0.0",
    "ruff>=0.9.7",
    "time-machine>=2.16.0",
    "vermin>=1.6.0",
]

[tool.pyright]
ignore = ["**/migrations", "**/*.ipynb"]
typeCheckingMode = "basic"
reportUnnecessaryTypeIgnoreComment = true

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "example_project.settings.development"
pythonpath = ["example_project"]
testpaths = ["adit_radis_shared/**/tests", "example_project/**/tests"]
python_files = ["tests.py", "test_*.py", "*_tests.py"]
log_cli = 0
log_cli_level = "INFO"
log_cli_format = "%(asctime)s [%(levelname)8s] %(message)s (%(filename)s:%(lineno)s)"
log_cli_date_format = "%Y-%m-%d %H:%M:%S"
markers = ["acceptance: mark a test as an acceptance test."]
filterwarnings = [
    # We already fixed this, so we only need to remove this ignore with next major version of factory boy
    "ignore:.*Factory._after_postgeneration will stop saving the instance:DeprecationWarning",
    "ignore:'cgi' is deprecated:DeprecationWarning",
    'ignore:.*use of fork\(\) may lead to deadlocks.*:DeprecationWarning',
    'ignore:.*Use timezone-aware objects to represent datetimes in UTC.*:DeprecationWarning',
]
timeout = 60

[tool.coverage.run]
branch = true
source = ["adit_radis_shared"]

[tool.coverage.report]
skip_empty = true
show_missing = true

[tool.ruff]
target-version = "py312"
exclude = ["migrations"]
line-length = 100
lint.select = ["E", "F", "I", "DJ"]

[tool.djlint]
profile = "django"
max_line_length = 120
ignore = "H021,H030,H031,T002"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
